test: use snapshots
