+++ src/index.js
@@ -1,6 +1,10 @@
 // @flow
 import createTransformer from './transform';
-import type { RollupPluginLux } from './interfaces';
+import type { TransformFunction } from './transform';
+
+export type RollupPluginLux = {
+  transform: TransformFunction;
+};
 
 export default function plugin(appPath: string): RollupPluginLux {
   return {
