--- src/interfaces.js
@@ -1,6 +1,0 @@
-// @flow
-import type { TransformFunction } from './transform';
-
-export type RollupPluginLux = {
-  transform: TransformFunction;
-};
