+++ src/test/__fixtures__/export-default-arrow-function.js
