+++ src/test/__fixtures__/export-default-expression-lux-sub-class.js
