+++ src/test/__fixtures__/export-default-function.js
