+++ src/test/__fixtures__/export-default-statement-generic-class.js
