+++ src/test/__fixtures__/export-default-statement-lux-sub-class.js
