+++ src/test/__fixtures__/named-export.js
