--- src/test/results/export-default-expression-lux-sub-class.js
@@ -1,7 +1,0 @@
-import { Model } from 'lux-framework';
-
-export default class User extends Model {
-
-}
-
-Object.defineProperty(User, 'name', { value: 'User' });
