--- src/test/results/export-default-statement-generic-class.js
@@ -1,9 +1,0 @@
-class Point {
-  constructor(x, y) {
-    Object.assign(this, { x, y });
-  }
-}
-
-Object.defineProperty(Point, 'name', { value: 'Point' });
-
-export default Point;
