--- src/test/results/export-default-statement-lux-sub-class.js
@@ -1,9 +1,0 @@
-import { Model } from 'lux-framework';
-
-class User extends Model {
-
-}
-
-Object.defineProperty(User, 'name', { value: 'User' });
-
-export default User;
