+++ src/transform/parse-source.js
@@ -4,7 +4,7 @@ import MagicString from 'magic-string';
 
 import { PARSER_OPTIONS } from '../constants';
 
-import type { TransformParams } from './interfaces';
+import type { TransformParams } from './index';
 
 export default function parseSource(src: string): TransformParams {
   return {
